pyproject.toml
[project]
name = "spring-financial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.8",
    "ipykernel>=6.30.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "xgboost>=3.0.2",
]
